Fix(snippet): Error that occurred when there were no linked resources in the document
When there are no linked resources in the document and the snippet getSelections is called, an error occurs due to an incorrectly composed query "sortby FIELD(...) in getSelections snippet.
Also, as far as I understand, it makes no sense to call the snippet getResources (pdoResources) if we already know that there are no linked resources. I suggest returning an empty result immediately if there are no linked resources in the document.